KL-163/refactor: move ProfileImage into profile
